// src/types/4337.ts
export interface UserOperationCallData {
	/* the target of the call */
	target: Address;
	/* the data passed to the target */
	data: Hex;
	/* the amount of native token to send to the target (default: 0) */
	value?: bigint;
}